move add_properties_to_request function into L8RequestObject as a method
